+++ search-films/src/components/SimpleMenuItem/SimpleMenuItem.props.ts
@@ -1,7 +1,7 @@
 import { HtmlHTMLAttributes, MouseEventHandler, ReactNode } from 'react';
 import { IconImage } from '../shared/MenuIconBuilder';
 
-export interface LinkedMenuItemProps extends HtmlHTMLAttributes<HTMLLIElement> {
+export interface SimpleMenuItemProps extends HtmlHTMLAttributes<HTMLLIElement> {
     children?: ReactNode;
 
     /**
@@ -10,19 +10,17 @@ export interface LinkedMenuItemProps extends HtmlHTMLAttributes<HTMLLIElement> {
     text: string;
 
     /**
-     * (optional)
-     * menu item link
-     */
-    linkUrl?: string;
-
-    /**
      * (optional) icon image
      */
     icon?: IconImage;
 
     /**
+     * (optional, byDefault=true) flag to handle clicks
+     */
+    canChoose?: boolean; 
+
+    /**
      * (optional, enabled if canChoose) function to handle clicks on menu button
      */
-    handleClicks?: MouseEventHandler<HTMLDivElement | HTMLLIElement>;
+    handleClicks?: MouseEventHandler<HTMLLIElement>;
 };
-