+++ search-films/src/main.tsx
@@ -3,36 +3,35 @@ import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { UserProvider } from './context/user.context';
 import './index.css';
+import MenuLayout from './layouts/MenuLayout';
 import ErrorPage from './pages/ErrorPage/index.js';
 import FavoritesPage from './pages/FavoritesPage/index.js';
 import { LoginPage, MainPage } from './pages/index.js';
-import MoviePage from './pages/MoviePage/index.js';
 
 const router = createBrowserRouter([
 	{
 		path: '/',
-		element: <MainPage/>,
+		element: <MenuLayout/>,
 		children: [
 			{
-				path: '/movies/:movieId',
-				element: <MoviePage/>
+				path: '/',
+				element: <MainPage/>
+			},
+			{
+				path: '/favorites',
+				element: <FavoritesPage/>
+			},
+			{
+				path: '/login',
+				element: <LoginPage/>
+			},
+			{
+				path: '*',
+				element: <ErrorPage/>
+
 			}
 		]
-	},
-	{
-		path: '/favorites',
-		element: <FavoritesPage/>
-	},
-	{
-		path: '/login',
-		element: <LoginPage/>
-	},
-	{
-		path: '*',
-		element: <ErrorPage/>
-
 	}
-    
 ]);
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(